+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.0
+
+- Rename class `Carol` -> `Pattern`
+- Type Declaration: Add documents.
+
 ## 0.1.0
 
 Initial Release.